+++ govtool/frontend/src/components/organisms/RegisterAsDRepSteps/RolesAndResponsibilities.tsx
@@ -20,7 +20,7 @@ export const RolesAndResponsibilities = ({
   const { t } = useTranslation();
   const { isMobile } = useScreenDimension();
 
-  const deposit = getItemFromLocalStorage(PROTOCOL_PARAMS_KEY);
+  const epochParams = getItemFromLocalStorage(PROTOCOL_PARAMS_KEY);
 
   const onClickContinue = () => setStep(2);
 
@@ -51,7 +51,7 @@ export const RolesAndResponsibilities = ({
             />,
           ]}
           i18nKey="registration.rolesAndResponsibilitiesDescription"
-          values={{ deposit: correctAdaFormat(deposit.drep_deposit) }}
+          values={{ deposit: correctAdaFormat(epochParams?.drep_deposit) }}
         />
       </Typography>
       <CenteredBoxBottomButtons